quiz true/false-show score,progress-MVC pattern
Quiz questions are in the QuizBrain.swift file and question is a data type which is created in Question.swift file's structure.
View controller uses the function and variables from brain page.
Questions are listed according to their question numbers, and the app will show progress and score. MVC pattern is used.